refactor: 필터 관련 컴포넌트 삭제 및 PostsTable에서 필터 상태 통합
