+++ src/features/filters/ui/FilterWrapper.tsx
@@ -11,5 +11,3 @@ export default function FilterWrapper() {
     </div>
   )
 }
-
-
