+++ src/features/filters/ui/TagFilter.tsx
@@ -1,15 +1,34 @@
-import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from "../../../../shared/ui"
+import { useEffect } from "react"
+import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from "../../../shared/ui"
 import { useShallow } from "zustand/shallow"
-import useFilterStore from "../../model/useFilterStore"
+import useFilterStore from "../model/useFilterStore"
+import { useTagsQuery } from "../../../entities/post/model/queries"
+
+interface Tag {
+  url: string
+  slug: string
+}
+
 export default function TagFilter() {
-  const { tags, selectedTag, setSelectedTag, setSkip } = useFilterStore(
+  const { selectedTag, setSelectedTag, setSkip, setTags } = useFilterStore(
     useShallow((state) => ({
       tags: state.tags,
       selectedTag: state.selectedTag,
       setSelectedTag: state.setSelectedTag,
       setSkip: state.setSkip,
+      setTags: state.setTags,
     })),
   )
+
+  // 태그 데이터를 직접 쿼리
+  const { data: tagsData } = useTagsQuery()
+
+  // 태그 데이터가 로드되면 상태 업데이트
+  useEffect(() => {
+    if (tagsData) {
+      setTags(tagsData)
+    }
+  }, [tagsData, setTags])
 
   return (
     <Select
@@ -24,7 +43,7 @@ export default function TagFilter() {
       </SelectTrigger>
       <SelectContent>
         <SelectItem value="all">모든 태그</SelectItem>
-        {tags.map((tag) => (
+        {tagsData?.map((tag: Tag) => (
           <SelectItem key={tag.url} value={tag.slug}>
             {tag.slug}
           </SelectItem>
