+++ src/widgets/PostDetailDialog/PostDetailDialog.tsx
@@ -1,12 +1,9 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../shared/ui/Dialog"
 import usePostsStore from "../../features/posts/model/usePostsStore"
 import useFilterStore from "../../features/filters/model/useFilterStore"
-import useCommentStore from "../../features/comment/model/useCommentStore"
 import CommentsList from "../../features/comment/ui/CommentsList"
 import highlightText from "../../shared/lib/util/highlightText"
 import { fetchComments } from "../../entities/comment/api/fetchComments"
-import { likeComment } from "../../entities/comment/api/likeComment"
-import { deleteComment } from "../../entities/comment/api/deleteComment"
 import { useShallow } from "zustand/shallow"
 import { useEffect } from "react"
 
@@ -25,17 +22,6 @@ const PostDetailDialog = () => {
     })),
   )
 
-  const { comments, setSelectedComment, setNewComment, setShowAddCommentDialog, setShowEditCommentDialog } =
-    useCommentStore(
-      useShallow((state) => ({
-        comments: state.comments,
-        setSelectedComment: state.setSelectedComment,
-        setNewComment: state.setNewComment,
-        setShowAddCommentDialog: state.setShowAddCommentDialog,
-        setShowEditCommentDialog: state.setShowEditCommentDialog,
-      })),
-    )
-
   // selectedPost가 변경될 때 댓글 가져오기
   useEffect(() => {
     if (selectedPost?.id) {
@@ -51,19 +37,7 @@ const PostDetailDialog = () => {
         </DialogHeader>
         <div className="space-y-4">
           <p>{highlightText(selectedPost?.body || "", searchQuery)}</p>
-          {selectedPost?.id && (
-            <CommentsList
-              postId={selectedPost.id}
-              comments={comments}
-              searchQuery={searchQuery}
-              setNewComment={setNewComment}
-              setShowAddCommentDialog={setShowAddCommentDialog}
-              likeComment={likeComment}
-              setSelectedComment={setSelectedComment}
-              setShowEditCommentDialog={setShowEditCommentDialog}
-              deleteComment={deleteComment}
-            />
-          )}
+          {selectedPost?.id && <CommentsList postId={selectedPost.id} />}
         </div>
       </DialogContent>
     </Dialog>
